Extract knowledge shared manager orchestration
